# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Learning Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students with optional filters
      parameters:
        - name: trainerId
          in: query
          required: false
          schema:
            type: integer
        - name: zoneId
          in: query
          required: false
          schema:
            type: integer
        - name: courseId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{id}/progress:
    get:
      operationId: getStudentProgress
      tags: [students]
      summary: Get complete progress for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgress"

  # ── COURSES ──────────────────────────────────────────────────────────────────
...
  /progress:
    get:
      operationId: listProgress
      tags: [progress]
      summary: List progress records with filters
      parameters:
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
        - name: trainerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of progress records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressRecord"
    post:
      operationId: submitProgress
      tags: [progress]
      summary: Submit or update a rating for a student module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "201":
          description: Progress submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressRecord"

  # ── AUDIT LOGS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
...
    Student:
      type: object
      required: [id, name, phone, courseId, courseName]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        timing:
          type: ["string", "null"]
        centreName:
          type: ["string", "null"]
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
    StudentInput:
      type: object
      required: [name, phone, courseId]
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: string
        courseId:
          type: integer
        timing:
          type: ["string", "null"]
        centreName:
          type: ["string", "null"]
        trainerId:
          type: ["integer", "null"]
    StudentDetail:
      type: object
      required: [id, name, phone, courseId, courseName]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        courseSyllabus:
          type: ["string", "null"]
        timing:
          type: ["string", "null"]
        centreName:
          type: ["string", "null"]
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
    ModuleProgressItem:
      type: object
      required: [moduleId, moduleName, rating]
      properties:
        moduleId:
          type: integer
        moduleName:
          type: string
        rating:
          type: ["string", "null"]
        comments:
          type: ["string", "null"]
        trainerId:
          type: ["integer", "null"]
        submittedAt:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
    ChapterProgressItem:
      type: object
      required: [chapterId, chapterName, modules]
      properties:
        chapterId:
          type: integer
        chapterName:
          type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleProgressItem"
    UnitProgressItem:
      type: object
      required: [unitId, unitName, chapters]
      properties:
        unitId:
          type: integer
        unitName:
          type: string
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/ChapterProgressItem"
    StudentBasicInfo:
      type: object
      required: [id, name, phone, courseName]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        courseName:
          type: string
        timing:
          type: ["string", "null"]
        centreName:
          type: ["string", "null"]
    StudentProgress:
      type: object
      required: [student, completionRate, completedModules, totalModules, units]
      properties:
        student:
          $ref: "#/components/schemas/StudentBasicInfo"
        completionRate:
          type: integer
        completedModules:
          type: integer
        totalModules:
          type: integer
        units:
          type: array
          items:
            $ref: "#/components/schemas/UnitProgressItem"

    # COURSE HIERARCHY
...
    ProgressRecord:
      type: object
      required: [id, studentId, studentName, moduleId, moduleName, chapterName, unitName, rating, submittedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        moduleId:
          type: integer
        moduleName:
          type: string
        chapterName:
          type: string
        unitName:
          type: string
        rating:
          type: string
          enum: [totally_dependent, physical_prompting, verbal_prompting, clueing, independent]
        comments:
          type: ["string", "null"]
          description: Optional trainer comments (max 500 chars)
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    ProgressInput:
      type: object
      required: [studentId, moduleId, rating]
      properties:
        studentId:
          type: integer
        moduleId:
          type: integer
        trainerId:
          type: ["integer", "null"]
        rating:
          type: string
          enum: [totally_dependent, physical_prompting, verbal_prompting, clueing, independent]
        comments:
          type: ["string", "null"]
          description: Optional comments (max 500 chars)

    # DASHBOARDS
...