# lib/api-zod/src/generated/types/chapterProgressItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleProgressItem } from './moduleProgressItem';

export interface ChapterProgressItem {
  chapterId: number;
  chapterName: string;
  modules: ModuleProgressItem[];
}