# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  phone: string;
  courseId: number;
  courseName: string;
  /** @nullable */
  timing?: string | null;
  /** @nullable */
  centreName?: string | null;
  /** @nullable */
  trainerId?: number | null;
  /** @nullable */
  trainerName?: string | null;
  /** @nullable */
  zoneId?: number | null;
  /** @nullable */
  zoneName?: string | null;
}